refactor: integrate lingui.js as lightweight i18n solution
